requirements.txt
click==8.1.7
PyYAML==6.0.2
python-docx==1.1.2
pypdf==4.3.1
regex==2024.9.11
rich==13.9.1
Flask==3.0.3
openai==1.43.0
